src/bid_system/bids_system/requirements.txt
Flask
pymongo
python-dotenv
dataclasses-json
nats-py
uuid